refactor: 프로젝트 섹션 애니메이션 등장 부분 변경
