+++ client/src/components/sections/ProjectsSection2.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef } from "react";
 import gsap from "gsap";
 import ScrollTrigger from "gsap/ScrollTrigger";
-import NoctaBackground from "../Background/noctaBackground";
+import NoctaBackground from "./components/noctaBackground";
 
 gsap.registerPlugin(ScrollTrigger);
 
@@ -16,13 +16,13 @@ export default function ProjectsSection2() {
       scrollTrigger: {
         trigger: sectionRef.current,
         start: "top center",
-        end: "center center",
+        end: "40% center",
         scrub: 1,
         markers: true,
         toggleActions: "play none none reverse",
       },
       opacity: 1,
-      y: 0,
+      x: 0,
       duration: 1,
     };
 
@@ -31,7 +31,7 @@ export default function ProjectsSection2() {
       {
         // Starting state
         opacity: 0,
-        y: 100,
+        x: -500,
         scale: 0.9,
       },
       {
@@ -46,17 +46,17 @@ export default function ProjectsSection2() {
   }, []);
 
   return (
-    <section ref={sectionRef} className="flex relative min-h-screen">
+    <section ref={sectionRef} className="flex relative justify-center">
       <NoctaBackground />
       <div className="absolute z-10 container mx-auto px-4 py-20">
         <div
           ref={projectCardRef}
           className="project-card backdrop-blur-md bg-white/10 rounded-2xl p-8 transform"
         >
-          <h3 className="text-2xl font-semibold mb-4 text-white">
+          <h3 className="text-2xl mb-4 text-white">
             실시간 동시편집 마크다운 에디터 Nocta
           </h3>
-          <p className="text-gray-200">프로젝트 설명이 들어갑니다.</p>
+          <p className="text-gray-200"> 기간 : 양랄ㅇ</p>
         </div>
       </div>
     </section>
